FEAT: add API connection

See merge request s12-fintech-finance-sub1/S12P21A207!35

+++ app/src/main/java/com/ilm/mulga/domain/model/MonthlyTransactionEntity.kt
@@ -1,7 +1,7 @@
 package com.ilm.mulga.domain.model
 
 data class MonthlyTransactionEntity(
-    val monthTotal: Long,
+    val monthTotal: Int,
     val year: Int,
     val month: Int,
     val dailySummaries: Map<Int, DailyTransactionSummaryEntity>,
